+++ src/utils/helpers/noteFetchers.ts
@@ -2,7 +2,10 @@ import fs from "fs";
 import { INote } from "../../interfaces/note";
 import path from "path";
 import { serialize } from "next-mdx-remote/serialize";
-import { Frontmatter } from "../../../pages/recipes/[slug]";
+import { GetStaticPropsContext } from "next";
+
+// TODO - for now, remove content prop used for md files, but cleanup later lol
+export type Frontmatter = Omit<INote, "content">;
 
 /**
  * Because I love relative linking, organizing things in a component-driven style, and categorizing my notes, I'm making my life harder than it needs to be :^)
@@ -11,7 +14,7 @@ import { Frontmatter } from "../../../pages/recipes/[slug]";
  * @param categories e.g. bread
  * @returns INote[]
  */
-export const getNotesForCategories = (
+export const getAllNotesForCategories = (
   baseFolder: string,
   categories: string[]
 ): INote[] => {
@@ -54,4 +57,49 @@ export const getNotesForCategories = (
   });
 
   return allNotes;
+};
+
+export const getNoteProps = async (
+  ctx: GetStaticPropsContext,
+  baseFolder: string,
+  categories: string[]
+) => {
+  const { slug } = ctx.params;
+
+  // Step 1 - get the mdx content
+  const source = fs.readFileSync(
+    path.join(
+      "_content/recipes/bread",
+      slug as string,
+      (slug + ".mdx") as string
+    ),
+    "utf8"
+  );
+
+  // Step 2 - reformat
+  const mdxSource = await serialize(source, { parseFrontmatter: true });
+
+  // Step 3 - return with proper types
+  return {
+    props: {
+      source: {
+        compiledSource: mdxSource.compiledSource,
+        scope: mdxSource.scope,
+        frontmatter: mdxSource.frontmatter as Frontmatter,
+      },
+      baseFolder: slug,
+    },
+  };
+};
+
+export const getNotePaths = (baseFolder: string, categories: string[]) => {
+  const files = fs.readdirSync("_content/recipes/bread");
+  return {
+    paths: files.map((file) => ({
+      params: {
+        slug: file,
+      },
+    })),
+    fallback: false,
+  };
 };
